+++ persistent-test/src/PersistentTestModelsImports.hs
@@ -1,0 +1,15 @@
+{-# language UndecidableInstances #-}
+
+-- | this just needs to compile
+module PersistentTestModelsImports where
+
+import Database.Persist.TH
+
+share [mkPersist sqlSettings] [persistUpperCase|
+
+User
+    name    String
+    age     Int
+    deriving Eq Show
+
+|]